refactor: rename groupings folder to clients
